fix(rig): harden source removal
