+++ src/core/rig/install.rs
@@ -8,7 +8,6 @@
 use crate::error::{Error, Result};
 use crate::{extension, git, paths};
 use serde::{Deserialize, Serialize};
-use std::collections::BTreeMap;
 use std::fs;
 use std::path::{Path, PathBuf};
 
@@ -44,63 +43,6 @@ pub struct RigSourceMetadata {
     pub linked: bool,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub source_revision: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RigSourceListResult {
-    pub sources: Vec<RigSourceGroup>,
-    pub invalid: Vec<InvalidRigSourceMetadata>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RigSourceGroup {
-    pub source: String,
-    pub package_path: String,
-    pub package_id: String,
-    pub linked: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub source_revision: Option<String>,
-    pub rigs: Vec<RigSourceRig>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RigSourceRig {
-    pub id: String,
-    pub rig_path: String,
-    pub config_path: String,
-    pub config_present: bool,
-    pub config_owned: bool,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct InvalidRigSourceMetadata {
-    pub id: String,
-    pub metadata_path: String,
-    pub error: String,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RigSourceRemoveResult {
-    pub selector: String,
-    pub source: RigSourceGroup,
-    pub removed: Vec<RemovedRigSourceRig>,
-    pub skipped: Vec<SkippedRigSourceRig>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub removed_package_path: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RemovedRigSourceRig {
-    pub id: String,
-    pub config_path: String,
-    pub metadata_path: String,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct SkippedRigSourceRig {
-    pub id: String,
-    pub config_path: String,
-    pub reason: String,
 }
 
 pub fn install(source: &str, id: Option<&str>, all: bool) -> Result<RigInstallResult> {
@@ -157,229 +99,6 @@ pub fn read_source_metadata(id: &str) -> Option<RigSourceMetadata> {
     let path = paths::rig_source_metadata(id).ok()?;
     let content = fs::read_to_string(path).ok()?;
     serde_json::from_str(&content).ok()
-}
-
-pub fn list_sources() -> Result<RigSourceListResult> {
-    read_source_entries().map(group_source_entries)
-}
-
-pub fn remove_source(selector: &str) -> Result<RigSourceRemoveResult> {
-    let list = list_sources()?;
-    let matches = list
-        .sources
-        .into_iter()
-        .filter(|source| source_matches(source, selector))
-        .collect::<Vec<_>>();
-
-    if matches.is_empty() {
-        return Err(Error::validation_invalid_argument(
-            "source",
-            format!("No installed rig source matches '{}'", selector),
-            Some(selector.to_string()),
-            None,
-        ));
-    }
-    if matches.len() > 1 {
-        let tried = matches
-            .iter()
-            .map(|source| source.package_id.clone())
-            .collect::<Vec<_>>();
-        return Err(Error::validation_invalid_argument(
-            "source",
-            format!(
-                "Selector '{}' matches multiple rig sources; use the full source or package path",
-                selector
-            ),
-            Some(selector.to_string()),
-            Some(tried),
-        ));
-    }
-
-    let source = matches.into_iter().next().expect("checked non-empty");
-    let mut removed = Vec::new();
-    let mut skipped = Vec::new();
-    for rig in &source.rigs {
-        let metadata_path = paths::rig_source_metadata(&rig.id)?;
-        let config_path = PathBuf::from(&rig.config_path);
-        if rig.config_present && rig.config_owned {
-            fs::remove_file(&config_path)
-                .map_err(|e| Error::internal_io(e.to_string(), Some("remove rig config".into())))?;
-            fs::remove_file(&metadata_path).map_err(|e| {
-                Error::internal_io(e.to_string(), Some("remove rig source metadata".into()))
-            })?;
-            removed.push(RemovedRigSourceRig {
-                id: rig.id.clone(),
-                config_path: rig.config_path.clone(),
-                metadata_path: metadata_path.to_string_lossy().to_string(),
-            });
-        } else if !rig.config_present {
-            fs::remove_file(&metadata_path).map_err(|e| {
-                Error::internal_io(e.to_string(), Some("remove rig source metadata".into()))
-            })?;
-            removed.push(RemovedRigSourceRig {
-                id: rig.id.clone(),
-                config_path: rig.config_path.clone(),
-                metadata_path: metadata_path.to_string_lossy().to_string(),
-            });
-        } else {
-            skipped.push(SkippedRigSourceRig {
-                id: rig.id.clone(),
-                config_path: rig.config_path.clone(),
-                reason: "config file no longer points at the recorded rig source".to_string(),
-            });
-        }
-    }
-
-    let removed_package_path = if !source.linked && skipped.is_empty() {
-        let package_path = PathBuf::from(&source.package_path);
-        if package_path.exists() {
-            fs::remove_dir_all(&package_path).map_err(|e| {
-                Error::internal_io(e.to_string(), Some("remove rig package".into()))
-            })?;
-            Some(source.package_path.clone())
-        } else {
-            None
-        }
-    } else {
-        None
-    };
-
-    Ok(RigSourceRemoveResult {
-        selector: selector.to_string(),
-        source,
-        removed,
-        skipped,
-        removed_package_path,
-    })
-}
-
-#[derive(Debug)]
-struct RigSourceEntry {
-    id: String,
-    metadata: RigSourceMetadata,
-}
-
-#[derive(Debug)]
-struct SourceEntries {
-    valid: Vec<RigSourceEntry>,
-    invalid: Vec<InvalidRigSourceMetadata>,
-}
-
-fn read_source_entries() -> Result<SourceEntries> {
-    let dir = paths::rig_sources()?;
-    if !dir.exists() {
-        return Ok(SourceEntries {
-            valid: Vec::new(),
-            invalid: Vec::new(),
-        });
-    }
-
-    let mut valid = Vec::new();
-    let mut invalid = Vec::new();
-    for entry in fs::read_dir(&dir)
-        .map_err(|e| Error::internal_io(e.to_string(), Some("read rig sources dir".into())))?
-    {
-        let entry = entry
-            .map_err(|e| Error::internal_io(e.to_string(), Some("read rig source entry".into())))?;
-        let path = entry.path();
-        if path.extension().and_then(|s| s.to_str()) != Some("json") {
-            continue;
-        }
-        let id = match path.file_stem().and_then(|s| s.to_str()) {
-            Some(id) => id.to_string(),
-            None => continue,
-        };
-        let content = match fs::read_to_string(&path) {
-            Ok(content) => content,
-            Err(err) => {
-                invalid.push(InvalidRigSourceMetadata {
-                    id,
-                    metadata_path: path.to_string_lossy().to_string(),
-                    error: err.to_string(),
-                });
-                continue;
-            }
-        };
-        match serde_json::from_str::<RigSourceMetadata>(&content) {
-            Ok(metadata) => valid.push(RigSourceEntry { id, metadata }),
-            Err(err) => invalid.push(InvalidRigSourceMetadata {
-                id,
-                metadata_path: path.to_string_lossy().to_string(),
-                error: err.to_string(),
-            }),
-        }
-    }
-
-    valid.sort_by(|a, b| a.id.cmp(&b.id));
-    invalid.sort_by(|a, b| a.id.cmp(&b.id));
-    Ok(SourceEntries { valid, invalid })
-}
-
-fn group_source_entries(entries: SourceEntries) -> RigSourceListResult {
-    let mut groups: BTreeMap<String, RigSourceGroup> = BTreeMap::new();
-    for entry in entries.valid {
-        let key = format!("{}\0{}", entry.metadata.source, entry.metadata.package_path);
-        let config_path = paths::rig_config(&entry.id).ok();
-        let config_present = config_path.as_ref().is_some_and(|path| path.exists());
-        let config_owned = config_path
-            .as_ref()
-            .is_some_and(|path| rig_config_matches_source(path, &entry.metadata.rig_path));
-
-        groups
-            .entry(key)
-            .or_insert_with(|| RigSourceGroup {
-                source: entry.metadata.source.clone(),
-                package_id: package_id_from_path(&entry.metadata.package_path),
-                package_path: entry.metadata.package_path.clone(),
-                linked: entry.metadata.linked,
-                source_revision: entry.metadata.source_revision.clone(),
-                rigs: Vec::new(),
-            })
-            .rigs
-            .push(RigSourceRig {
-                id: entry.id,
-                rig_path: entry.metadata.rig_path,
-                config_path: config_path
-                    .map(|path| path.to_string_lossy().to_string())
-                    .unwrap_or_default(),
-                config_present,
-                config_owned,
-            });
-    }
-
-    let mut sources = groups.into_values().collect::<Vec<_>>();
-    for source in &mut sources {
-        source.rigs.sort_by(|a, b| a.id.cmp(&b.id));
-    }
-
-    RigSourceListResult {
-        sources,
-        invalid: entries.invalid,
-    }
-}
-
-fn source_matches(source: &RigSourceGroup, selector: &str) -> bool {
-    source.source == selector || source.package_path == selector || source.package_id == selector
-}
-
-fn package_id_from_path(path: &str) -> String {
-    Path::new(path)
-        .file_name()
-        .and_then(|name| name.to_str())
-        .unwrap_or(path)
-        .to_string()
-}
-
-fn rig_config_matches_source(config_path: &Path, rig_path: &str) -> bool {
-    let rig_path = Path::new(rig_path);
-    if fs::read_link(config_path).is_ok_and(|target| target == rig_path) {
-        return true;
-    }
-
-    match (config_path.canonicalize(), rig_path.canonicalize()) {
-        (Ok(config), Ok(rig)) => config == rig,
-        _ => false,
-    }
 }
 
 struct PreparedSource {
